Fix QR codes for phone #'s with full int'l prefix
This one has bugged me for a long time but only now I got so irritated to stop and think of a solution.

Problem: when a customer inputs their phone number with the international dial code in front (e.g .+359888123456), PHP will replace that plus sign with a space because the plus symbol is reserved (RFC 2396) and cannot be used within GET queries. Because of this, the QR for the phone number will present as tel:359888123456 which most Android dialers will not recognize as a phone number (IDK about iOS)

Solution: I have wrapped the $telephone variable within an urlencode() which makes sure that the plus sign will be properly translated as %2B and passed onto the QR code generation function.